report_engine/requirements.txt
yfinance
feedparser
google-genai
python-dotenv
pandas
pandas_market_calendars
requests
